% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa5Icon.R
\name{fa5Icon}
\alias{fa5Icon}
\title{Make a map marker with a FontAwesome 5 icon}
\usage{
fa5Icon(name, color = "blue")
}
\arguments{
\item{name}{Name of the icon ("anchor")}

\item{color}{One of 11 colors}
}
\description{
Makes an icon for use in `leaflet` with the `makeIcon` function.
}
