% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_tbl.R
\docType{data}
\name{fa_tbl}
\alias{fa_tbl}
\title{Complete list of FontAwesome icons (5+)}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{name}{short name of the icon, "address-book"}
  \item{style}{brands, solid or regular}
  \item{full_name}{full name of the icon, "fas fa-address-book"}
}
}
\usage{
fa_tbl
}
\description{
An up to date list of all available FontAwesome Icons, most recent version (5+ currently).
}
\keyword{datasets}
