% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvk_icon_mapping.R
\docType{data}
\name{kvk_icon_mapping}
\alias{kvk_icon_mapping}
\title{Mapping of KvK branche code (sbi) to fontawesome icons}
\format{
A data frame with 92 rows and 2 variables:
\describe{
  \item{sbi}{KvK branche code}
  \item{fontawesomename}{name of fontawesome (v5+) icon}
}
}
\usage{
kvk_icon_mapping
}
\description{
Mapping of KvK branche code (sbi) to fontawesome icons
}
\keyword{datasets}
