% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvkIcons.R
\name{kvkIcons}
\alias{kvkIcons}
\title{Make map markers for KvK branches}
\usage{
kvkIcons(
  x,
  color = c("blue", "darkblue", "darkgreen", "darkpurple", "darkred", "green", "navy",
    "orange", "pink", "purple", "red"),
  names_only = FALSE
)
}
\arguments{
\item{x}{Vector of SBI codes (e.g. "01.11.1","6723","6"), with or without periods and spaces.}

\item{color}{Color of the map marker (one of a number of choices, not free to set).}

\item{names_only}{If TRUE, returns only the fontawesome name of the icon (for testing), otherwise (an)
icon(s) made with leaflet::makeIcon}
}
\description{
Make map markers for KvK branches
}
